Added a class on the project to restore the app bofore every test
